John's definitive 2 models (minus tensors) + tests of the model
Uploading John's two models he was not able to include into Github. These models do not have safetensors due to LFS limits. This is okay, considering the models themselves are unlikely to be used for the actual goal of the research (model 5 Roberta is likely the best).

Also included tests for John's models and included his documentation of how to use the golden_dataset app